Trim trailing slash from URI path in Apache config
This allows Apache to answer requests for the bare /phpipam URI, which
may be desired by some configs to avoid the need to add an extra
redirect. This seems to be handled by PHPIPAM correctly, although it
should be noted that the trailing slash is necessary still in both
config.php and .htaccess.