doc: Non-standard attribute caveat and release fetching info
Types not standard to the Chef server "should" be avoided, as they will
not save to the node properly, although they still may be useful (Regexp
literals seem to render OK as long as you use a wrapper cookbook, so
that's good, but they save to the Chef server as strings).

Also added some bits on how to add the cookbook to your projects since
it currently is not pushed to the Chef server.