Add phpipam-sdk-go acceptance test suite provisioning support
This adds a phpipam-sdk-go-testacc test suite to test-kitchen, and a
cookbook that loads in some schema changes that reconfigure the Admin
account's password, removes the passwordChange flag, configures API
access, and adds custom fields for addresses, subnets, and vlans.

These steps configure an instance that can be used for acceptance
testing the phpipam-sdk-go suite, and possibly also the
terraform-provider-phpipam plugin.